Organise the class into duke package and their subpackages
