+++ data/task.txt
@@ -1,3 +1,6 @@
 T | 1 | 123
-T | 0 | 12345
+T | 1 | 12345
 D | 1 | slap ym | Dec 12 2020
+T | 0 | hi hi
+T | 1 | cry
+T | 0 | test
