+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.ui.Storage;
+import duke.ui.TaskList;
+import duke.ui.Ui;
+
 public class ExitCommand extends Command { //ExitCommand to handle the shutting down of bots
 
     public ExitCommand() {
