+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Deadline extends Task{
     protected String dueDate;
