+++ src/main/java/duke/task/Event.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Event extends Task{
     protected String eventTime;
