+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Task { //Base Task Class
     private final String name;
     protected boolean isDone;
