+++ src/main/java/duke/ui/Storage.java
@@ -1,3 +1,5 @@
+package duke.ui;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Path;
