+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ Hello from
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
 
-Hello! I'm Duke
+Hello! I'm duke.Duke
 What can I do for you?
 
 _______________
